# frontend/api-contracts/openapi/client-service.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "OpenAPI definition",
                 "version":  "v0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8200/client-service",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
...
                  "/api/verificationTokens/{verificationTokenId}":  {
                                                                        "get":  {
                                                                                    "tags":  [
                                                                                                 "verification-token-resource"
                                                                                             ],
                                                                                    "operationId":  "findById",
                                                                                    "parameters":  [
                                                                                                       {
                                                                                                           "name":  "verificationTokenId",
                                                                                                           "in":  "path",
                                                                                                           "required":  true,
                                                                                                           "schema":  {
                                                                                                                          "type":  "string"
                                                                                                                      }
                                                                                                       }
                                                                                                   ],
                                                                                    "responses":  {
                                                                                                      "200":  {
                                                                                                                  "description":  "OK",
                                                                                                                  "content":  {
                                                                                                                                  "*/*":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/VerificationTokenDto"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                              }
                                                                                                              }
                                                                                                  }
                                                                                },
                                                                        "put":  {
                                                                                    "tags":  [
                                                                                                 "verification-token-resource"
                                                                                             ],
                                                                                    "operationId":  "update_1",
                                                                                    "parameters":  [
                                                                                                       {
                                                                                                           "name":  "verificationTokenId",
                                                                                                           "in":  "path",
                                                                                                           "required":  true,
                                                                                                           "schema":  {
                                                                                                                          "type":  "string"
                                                                                                                      }
                                                                                                       }
                                                                                                   ],
                                                                                    "requestBody":  {
                                                                                                        "content":  {
                                                                                                                        "application/json":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/VerificationTokenDto"
                                                                                                                                                            }
                                                                                                                                             }
                                                                                                                    },
                                                                                                        "required":  true
                                                                                                    },
                                                                                    "responses":  {
                                                                                                      "200":  {
                                                                                                                  "description":  "OK",
                                                                                                                  "content":  {
                                                                                                                                  "*/*":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/VerificationTokenDto"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                              }
                                                                                                              }
                                                                                                  }
                                                                                },
                                                                        "delete":  {
                                                                                       "tags":  [
                                                                                                    "verification-token-resource"
                                                                                                ],
                                                                                       "operationId":  "deleteById",
                                                                                       "parameters":  [
                                                                                                          {
                                                                                                              "name":  "verificationTokenId",
                                                                                                              "in":  "path",
                                                                                                              "required":  true,
                                                                                                              "schema":  {
                                                                                                                             "type":  "string"
                                                                                                                         }
                                                                                                          }
                                                                                                      ],
                                                                                       "responses":  {
                                                                                                         "200":  {
                                                                                                                     "description":  "OK",
                                                                                                                     "content":  {
                                                                                                                                     "*/*":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "type":  "boolean"
                                                                                                                                                            }
                                                                                                                                             }
                                                                                                                                 }
                                                                                                                 }
                                                                                                     }
                                                                                   }
                                                                    },
...
                  "/api/users/{userId}":  {
                                              "get":  {
                                                          "tags":  [
                                                                       "user-resource"
                                                                   ],
                                                          "operationId":  "findById_1",
                                                          "parameters":  [
                                                                             {
                                                                                 "name":  "userId",
                                                                                 "in":  "path",
                                                                                 "required":  true,
                                                                                 "schema":  {
                                                                                                "type":  "string"
                                                                                            }
                                                                             }
                                                                         ],
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/UserDto"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      },
                                              "put":  {
                                                          "tags":  [
                                                                       "user-resource"
                                                                   ],
                                                          "operationId":  "update_3",
                                                          "parameters":  [
                                                                             {
                                                                                 "name":  "userId",
                                                                                 "in":  "path",
                                                                                 "required":  true,
                                                                                 "schema":  {
                                                                                                "type":  "string"
                                                                                            }
                                                                             }
                                                                         ],
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/UserDto"
                                                                                                                                  }
                                                                                                                   }
                                                                                          },
                                                                              "required":  true
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/UserDto"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      },
                                              "delete":  {
                                                             "tags":  [
                                                                          "user-resource"
                                                                      ],
                                                             "operationId":  "deleteById_1",
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "userId",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "string"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "boolean"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           }
                                                         }
                                          },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "UserDto":  {
                                                       "type":  "object",
                                                       "properties":  {
                                                                          "userId":  {
                                                                                         "type":  "integer",
                                                                                         "format":  "int32"
                                                                                     },
                                                                          "firstName":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                          "lastName":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                          "imageUrl":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                          "email":  {
                                                                                        "type":  "string"
                                                                                    },
                                                                          "phone":  {
                                                                                        "type":  "string"
                                                                                    },
                                                                          "addressDtos":  {
                                                                                              "uniqueItems":  true,
                                                                                              "type":  "array",
                                                                                              "items":  {
                                                                                                            "$ref":  "#/components/schemas/AddressDto"
                                                                                                        }
                                                                                          },
                                                                          "credential":  {
                                                                                             "$ref":  "#/components/schemas/CredentialDto"
                                                                                         }
                                                                      }
                                                   },
                                       "VerificationTokenDto":  {
                                                                    "type":  "object",
                                                                    "properties":  {
                                                                                       "verificationTokenId":  {
                                                                                                                   "type":  "integer",
                                                                                                                   "format":  "int32"
                                                                                                               },
                                                                                       "token":  {
                                                                                                     "type":  "string"
                                                                                                 },
                                                                                       "expireDate":  {
                                                                                                          "type":  "string",
                                                                                                          "format":  "date"
                                                                                                      },
                                                                                       "credential":  {
                                                                                                          "$ref":  "#/components/schemas/CredentialDto"
                                                                                                      }
                                                                                   }
                                                                },
...
                                   }
                   }
}